+++ lib/data/data_sources/remote/abstract/skeleton_data_source.dart
@@ -1,0 +1,6 @@
+import 'package:flutter_base_rootstrap/utils/resource.dart';
+import 'package:flutter_base_rootstrap/utils/custom_exception.dart';
+
+abstract class SkeletonDataSource {
+  Future<Resource<bool, CustomException>> checkServerStatus();
+}
