+++ lib/data/repositories/skeleton_repository_impl.dart
@@ -1,0 +1,15 @@
+import 'package:flutter_base_rootstrap/skeleton/data/data_sources/remote/abstract/skeleton_data_source.dart';
+import 'package:flutter_base_rootstrap/skeleton/domain/repositories/skeleton_repository.dart';
+import 'package:flutter_base_rootstrap/utils/custom_exception.dart';
+import 'package:flutter_base_rootstrap/utils/resource.dart';
+
+class SkeletonRepositoryImpl extends SkeletonRepository {
+  final SkeletonDataSource _skeletonDataSource;
+
+  SkeletonRepositoryImpl(this._skeletonDataSource);
+
+  @override
+  Future<Resource<bool, CustomException>> checkServerStatus() async {
+    return await _skeletonDataSource.checkServerStatus();
+  }
+}
