+++ lib/di_init.dart
@@ -1,3 +1,5 @@
+import 'package:flutter_base_rootstrap/data/data_sources/local/abstract/preferences.dart';
+import 'package:flutter_base_rootstrap/data/data_sources/local/concrete/preferences.dart';
 import 'package:flutter_base_rootstrap/devices/permissions/abstract/permission_manager.dart';
 import 'package:flutter_base_rootstrap/devices/permissions/concrete/mobile/_permissions_android.dart';
 import 'package:flutter_base_rootstrap/devices/permissions/concrete/mobile/_permissions_ios.dart';
@@ -8,14 +10,7 @@ import 'package:flutter_base_rootstrap/devices/platform/concrete/_app_platform_i
     if (dart.library.io) 'package:flutter_base_rootstrap/devices/platform/concrete/mobile_desk/_app_platform_impl.dart'
     if (dart.library.html) 'package:flutter_base_rootstrap/devices/platform/concrete/web/_app_platform_impl.dart';
 import 'package:flutter_base_rootstrap/devices/platform/concrete/_platform_info_impl.dart';
-import 'package:flutter_base_rootstrap/domain/services/abstract/service_class.dart';
-import 'package:flutter_base_rootstrap/domain/services/concrete/service_class.dart';
-import 'package:flutter_base_rootstrap/repository/abstract/repository_class.dart';
-import 'package:flutter_base_rootstrap/repository/concrete/repository_class.dart';
-import 'package:flutter_base_rootstrap/repository/data_source/local/abstract/preferences.dart';
-import 'package:flutter_base_rootstrap/repository/data_source/local/concrete/preferences.dart';
-import 'package:flutter_base_rootstrap/repository/data_source/remote/abstract/remote_ds_class.dart';
-import 'package:flutter_base_rootstrap/repository/data_source/remote/concret/remote_ds_class.dart';
+
 import 'package:flutter_base_rootstrap/skeleton/data/data_sources/remote/abstract/skeleton_data_source.dart';
 import 'package:flutter_base_rootstrap/skeleton/data/data_sources/remote/concrete/skeleton_data_source_impl.dart';
 import 'package:flutter_base_rootstrap/skeleton/data/repositories/skeleton_repository_impl.dart';
@@ -36,17 +31,11 @@ import 'package:shared_preferences/shared_preferences.dart';
 // **//
 Future<void> initialize() async {
   final pref = await SharedPreferences.getInstance();
+
   getIt.registerSingleton<SharedPreferences>(pref);
   getIt.registerSingleton<Preferences>(PreferencesImpl(getIt()));
-  getIt.registerSingleton<RemoteDsExample>(RemoteDsExampleImpl());
   getIt.registerSingleton<AppPlatform>(AppPlatformImpl());
   getIt.registerLazySingleton<PlatformInfo>(() => PlatformInfoImpl(getIt()));
-  getIt.registerFactory<RepositoryExample>(
-    () => (RepositoryExampleImpl(getIt())),
-  );
-  getIt.registerLazySingleton<ServiceExample>(
-    () => ServiceExampleImpl(getIt()),
-  );
   getIt.registerLazySingleton<PermissionManager>(() {
     final platforms = getIt<AppPlatform>();
     if (platforms.isWeb) {
