+++ lib/domain/repositories/skeleton_repository.dart
@@ -1,0 +1,6 @@
+import 'package:flutter_base_rootstrap/utils/custom_exception.dart';
+import 'package:flutter_base_rootstrap/utils/resource.dart';
+
+abstract class SkeletonRepository {
+  Future<Resource<bool, CustomException>> checkServerStatus();
+}
