+++ lib/presentation/resources/resources.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_base_rootstrap/presentation/themes/spacing.dart';
+import 'package:flutter_svg/svg.dart';
+
+part 'dim.dart';
+
+part 'images.dart';
+
+extension SpacingOnWidget on Widget {
+  Spacing get spacing => Spacing();
+}
+
+extension SpacingOnStateWidget on State {
+  Spacing get spacing => Spacing();
+}
