+++ lib/presentation/ui/pages/skeleton/skeleton_view.dart
