+++ lib/presentation/ui/pages/skeleton/widgets/custom_widget.dart
@@ -1,0 +1,1 @@
+//Add here custom widgets needed
