+++ lib/presentation/ui/widgets/README.md
@@ -1,0 +1,3 @@
+### Custom Widgets
+
+- Custom Widgets used application wide
