--- lib/repository/abstract/repository_class.dart
@@ -1,3 +1,0 @@
-abstract class RepositoryExample {
-  Future<bool> isServerOnline();
-}
