--- lib/repository/data_source/local/abstract/preferences.dart
@@ -1,15 +1,0 @@
-import 'package:flutter_base_rootstrap/presenter/resources/locale/localize.dart';
-
-abstract class Preferences {
-  Lang get appLang;
-
-  bool get cookiesAllowed;
-
-  set appLang(Lang appLang);
-
-  set cookiesAllowed(bool cookiesAllowed);
-
-  set secureHeaders(Map<String, String> headers);
-
-  Map<String, String> get secureHeaders;
-}
